Nicholas Cvitak: 	- Fixed 32bit -> 64bit serialization bug. 	- All values are now stored as 32bit in PXPTIF

+++ FBTraining/Live2BenchNative/PXPTIFAction.h
@@ -14,9 +14,9 @@
 
 /// Data structure used to represent a point in a PXPTIF.
 typedef struct PXPTIFPoint {
-    double time;
+    float time;
     struct {
-        uint32_t x, y;
+        uint16_t x, y;
     } position;
 } PXPTIFPoint;
 
@@ -24,7 +24,7 @@ typedef struct PXPTIFPoint {
 typedef struct PXPTIFAction *PXPTIFActionRef;
 
 /// Creates an action.
-PXPTIFActionRef __nonnull PXPTIFActionCreate(uint32_t type, PXPTIFColor color, float width, PXPTIFPoint *__nullable points, uint64_t n_points);
+PXPTIFActionRef __nonnull PXPTIFActionCreate(uint32_t type, PXPTIFColor color, float width, PXPTIFPoint *__nullable points, uint32_t n_points);
 
 /// Destroys an action.
 void PXPTIFActionDestroy(PXPTIFActionRef __nonnull action);
@@ -42,12 +42,12 @@ PXPTIFColor PXPTIFActionGetColor(PXPTIFActionRef __nonnull action);
 float PXPTIFActionGetWidth(PXPTIFActionRef __nonnull action);
 
 /// Gets the number of points contained within an action.
-uint64_t PXPTIFActionGetPointCount(PXPTIFActionRef __nonnull action);
+uint32_t PXPTIFActionGetPointCount(PXPTIFActionRef __nonnull action);
 
 /// Gets the points contained within an action.
 const PXPTIFPoint *__nullable PXPTIFActionGetPoints(PXPTIFActionRef __nonnull action);
 
 /// Sets the size in bytes of the entire action data structure.
-uint64_t PXPTIFActionGetSize(PXPTIFActionRef __nonnull action);
+uint32_t PXPTIFActionGetSize(PXPTIFActionRef __nonnull action);
 
 #endif /* PXPTIFAction_c */
