+++ FBTraining/Live2BenchNative/PXPTIFByteBuffer.h
@@ -9,7 +9,7 @@
 #ifndef PXPTIFByteBuffer_c
 #define PXPTIFByteBuffer_c
 
-#include <inttypes.h>
+#include <stddef.h>
 
 /// Utility data structure used to format PXPTIF data in memory.
 typedef struct PXPTIFByteBuffer *PXPTIFByteBufferRef;
@@ -21,22 +21,22 @@ PXPTIFByteBufferRef __nonnull PXPTIFByteBufferCreate(void);
 void PXPTIFByteBufferDestroy(PXPTIFByteBufferRef __nonnull byteBuffer);
 
 /// Gets the size of the ByteBuffer's internal buffer.
-uint64_t PXPTIFByteBufferGetSize(PXPTIFByteBufferRef __nonnull byteBuffer);
+size_t PXPTIFByteBufferGetSize(PXPTIFByteBufferRef __nonnull byteBuffer);
 
 /// Gets the ByteBuffer's internal buffer.
 const void *__nullable PXPTIFByteBufferGetBuffer(PXPTIFByteBufferRef __nonnull byteBuffer);
 
 /// Gets the current read/write position in the ByteBuffer.
-uint64_t PXPTIFByteBufferGetPosition(PXPTIFByteBufferRef __nonnull byteBuffer);
+size_t PXPTIFByteBufferGetPosition(PXPTIFByteBufferRef __nonnull byteBuffer);
 
 /// Attempts to set the ByteBuffer's read/write position, and returns the position it went to.
-uint64_t PXPTIFByteBufferSetPosition(PXPTIFByteBufferRef __nonnull byteBuffer, uint64_t position);
+size_t PXPTIFByteBufferSetPosition(PXPTIFByteBufferRef __nonnull byteBuffer, size_t position);
 
 /// Writes 'size' many 'bytes' to the ByteBuffer at the current position, and moves the position forward by 'size'.
-void PXPTIFByteBufferWriteBytes(PXPTIFByteBufferRef __nonnull byteBuffer, const void *__nullable bytes, uint64_t size);
+void PXPTIFByteBufferWriteBytes(PXPTIFByteBufferRef __nonnull byteBuffer, const void *__nullable bytes, size_t size);
 
 /// Read's 'size' many 'bytes' from the ByteBuffer at the current position, and move the position forward by 'size'
-void PXPTIFByteBufferReadBytes(PXPTIFByteBufferRef __nonnull byteBuffer, void *__nullable bytes, uint64_t size);
+void PXPTIFByteBufferReadBytes(PXPTIFByteBufferRef __nonnull byteBuffer, void *__nullable bytes, size_t size);
